+++ backend/data/event.data.js
@@ -241,10 +241,11 @@ class Event {
         startingHour,
         endingHour,
         idEmployee,
+        workToDo,
         callback
     ) {
         const query =
-            "INSERT INTO events (title, description, status, is_planned, type, id_client, id_address, starting_date, starting_hour, ending_hour, id_employee) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11) RETURNING *";
+            "INSERT INTO events (title, description, status, is_planned, type, id_client, id_address, starting_date, starting_hour, ending_hour, id_employee, work_to_do) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12) RETURNING *";
         const values = [
             title,
             description,
@@ -257,6 +258,7 @@ class Event {
             startingHour,
             endingHour,
             idEmployee,
+            workToDo,
         ];
 
         pool.query(query, values, (error, result) => {
@@ -277,7 +279,8 @@ class Event {
                 row.starting_date,
                 row.starting_hour,
                 row.ending_hour,
-                row.id_employee
+                row.id_employee,
+                row.work_to_do
             );
 
             callback(null, newEvent);